+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Le format est basé sur [Tenez un Changelog](https://keepachangelog.com/fr/1.0.0
 ## [À venir dans la prochaine version]
 ### Ajouté
 - Permet de joindre un fichier à un fait marquant
+- Commande pour envoyer un mail aux utilisateurs pour leur rappeller de saisir leurs temps
 ### Modifié
 - Refactorise l'entité FichierProjet dans Fichier pour être plus flexible sur les fichiers
 - Utilise flysystem comme couche d'abstraction du système de fichier